Fix the rust build
This brings over all of the updater rust code from old_repo, except I deleted the "publisher" package which is now covered in the shorebird cli and I used the updated "shorebird_code_push_updater" package as "library".

I restored building with the rust and cdylib variants and kept the UPDATER_DEMO docs (we'll delete them soon but they still have some build instructions).  I edited down the README a bit since some of the instructions were old.

Once this is checked in, I'll work with felix to organize the code more aligned with his vision for the repo layout.