Add two more tests
Fixed report_successful_launch to not crash with no current patch.
Also fixed current_patch to not crash in the case of an invalid state.
Also fixed a potential crash in download_file_to_path but I don't have a mocking setup for the network yet to be able to test.