Add support for diff patches
This adds support for zstd-compressed patch files.

I ended up needing to change the engine code as well (it was sending the wrong path for the base libapp.so), so I also changed the signature of shorebird_active_version to shorebird_active_patch_number to match our modern wording.  It's still a char* for now.

The compression comes from zstd, through comde (compression/decompression) which is a simple wrapper library around a few rust compression libraries.  We could drop the intermediate layer once we're sure which compression algorithm we're using.

Fixed dont to do_not to appease cspell.

Added a new (optional) bool in the server Patch response "is_diff" which controls whether this behavior is turned on or not.

I've tested by manually submitting my own diff that this flow does work correctly, but we'll need to change the cli and server to support it by default.

Fixed a couple places we were using PathBuf instead of Path.  Turns out PathBuf is to String where Path& is to str&, so using Path& instead of PathBuf& where possible.