Teach rust side to validate hashes when installing
Patches which are sent down with a valid hash have that hash checked during install.

We could also (later) choose to check on every boot or otherwise validate the integrity of the cache.

Confirmed manually through deploys (and unittests) that both hashes are ignored when invalid (e.g. "#") and validated when valid.