chore(shorebird_cli): add the concept of an extension command
An extension command follows the Git model of CLI extension where any command unknown to the git CLI that matches the pattern git-<command> and is executable can be used to extend the Git command set.

For example:

git lfs

If the 'lfs' command is not implemented within Git's internal commands then the Git command line parser would check to see if 'git-lfs' exists and is executable.  If so, then the Git CLI passes execution to that script.