test: re-write cutler to be more testable
I removed the 'rebase' command for now since it was dead.
'rebase' should be re-written later to just print the commands
needed since having it actually drive the rebase was a dream.
I separated "Repo" from "Checkout" and removed the global config object.
I moved to using a global logger and package scoped to match
other packages.
I added a bunch of tests.  This still doesn't have 100% coverage
but it should
be 100% testable now.