feat: support field_table data in link supplement
Also refactored link_supplement code to be shared betwen
ios_patcher and ios_framework_patcher

I removed the behavior of the `shorebird` tool failing when the
link supplement was missing a file.  gen_snapshot or link will
fail if the files are missing no need to add extra complexity at
this layer.